Refactored Google Maps functions and moved it into Component level with props to RenderMap.js
