+++ .gitignore
@@ -25,3 +25,6 @@
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
+
+#Ignore Test files
+.AppTestIan.js
