+++ package.json
@@ -9,7 +9,6 @@
     "@mui/material": "^5.15.0",
     "@react-google-maps/api": "^2.19.2",
     "axios": "^1.6.2",
-    "@react-google-maps/api": "^2.19.2",
     "cors": "^2.8.5",
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
