+++ src/index.js
@@ -1,6 +1,7 @@
 import React from "react";
 import { createRoot } from "react-dom/client";
-import AppTestIan from "./AppTestIan";
+//import AppTestIan from "./AppTestIan";
+import App from "./App";
 
 const rootElement = document.getElementById("root");
-createRoot(rootElement).render(<AppTestIan />);
+createRoot(rootElement).render(<App />);
